backend/requirements.txt
# -------------------------------
# Web 框架 & 服务器
# -------------------------------

fastapi>=0.104.0
# FastAPI：后端主框架，用于构建 API（高性能、类型友好、自动生成 Swagger）

uvicorn[standard]>=0.24.0
# ASGI 服务器：运行 FastAPI
# [standard] 包含 uvloop、watchfiles、httptools 等优化组件


# -------------------------------
# 数据库相关
# -------------------------------

sqlalchemy>=2.0.0
# ORM 框架：用于定义模型、操作数据库（兼容 PostgreSQL / MySQL 等）

psycopg2-binary>=2.9.0
# PostgreSQL 官方驱动（binary 版本，免编译安装）

alembic>=1.12.0
# 数据库迁移工具


# -------------------------------
# 配置与环境变量
# -------------------------------

python-dotenv>=1.0.0
pydantic-settings>=2.0.0
# 环境变量管理，支持 .env 文件


# -------------------------------
# 数据校验 & 序列化
# -------------------------------

pydantic>=2.0.0
# 数据校验与序列化
# FastAPI 底层依赖，用于 Request / Response Schema


# -------------------------------
# 认证 & 安全
# -------------------------------

python-jose[cryptography]>=3.3.0
# JWT 生成与解析
# 用于 Access Token / Refresh Token 认证

passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
# 密码加密与校验
# bcrypt 是业界标准密码哈希算法（不可逆）


# -------------------------------
# 文件上传 / 表单支持
# -------------------------------

python-multipart>=0.0.9
# 处理 multipart/form-data
# 支持文件上传和表单数据


# -------------------------------
# 时区处理
# -------------------------------

pytz>=2023.3
# 时区处理库
# 用于时区转换和时区列表获取


# -------------------------------
# 开发工具
# -------------------------------

black>=23.0.0
# 代码格式化工具

flake8>=6.0.0
# 代码检查工具

pytest>=7.0.0
# 测试框架


# -------------------------------
# 其他实用工具
# -------------------------------

requests>=2.31.0
# HTTP 客户端

python-dateutil>=2.8.0
# 日期处理工具


# -------------------------------
# 日志相关
# -------------------------------

loguru>=0.7.0
# 现代化的日志库

